Seccion 6 - 65 - Se crea setting.service - Se encapsula la logica de el sethings de tema en el servicio
