+++ adminpro/src/app/pages/account-settings/account-settings.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { SettingsService } from 'src/app/services/settings.service';
 
 @Component({
   selector: 'app-account-settings',
@@ -8,32 +9,18 @@ import { Component, OnInit } from '@angular/core';
 })
 export class AccountSettingsComponent implements OnInit {
   
-  public linkTheme = document.querySelector('#theme');
   public links!: NodeListOf<Element>;
 
-
-  constructor() { }
+  constructor(private settingService: SettingsService) { }
 
   ngOnInit(): void {
-    this.links = document.querySelectorAll('.selector')
-    this.checkCurrentTheme();
+    this.links = document.querySelectorAll('.selector');
+    this.settingService.checkCurrentTheme(this.links);
   }
-
+  
   changeTheme( theme: string) {
-    const url = `./assets/css/colors/${theme}.css`
-    this.linkTheme?.setAttribute('href', url);
-    localStorage.setItem('theme', url);
-    this.checkCurrentTheme();
-  }
-
-  checkCurrentTheme() {
-    this.links.forEach(element => {
-      element.classList.remove('working')
-      const btnTheme = element.getAttribute('data-theme');
-      const btnThemeUrl = `./assets/css/colors/${btnTheme}.css`;
-      const currentTheme = this.linkTheme?.getAttribute('href');
-      if(btnThemeUrl === currentTheme) element.classList.add('working');
-    });
+    this.settingService.changeTheme(theme);
+    this.settingService.checkCurrentTheme(this.links);
   }
 
 }
