+++ adminpro/src/app/pages/pages.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { SettingsService } from '../services/settings.service';
 
 @Component({
   selector: 'app-pages',
@@ -8,13 +9,11 @@ import { Component, OnInit } from '@angular/core';
 })
 export class PagesComponent implements OnInit {
 
-  public linkTheme = document.querySelector('#theme');
 
-  constructor() { }
+  constructor(private settingService: SettingsService) { }
 
   ngOnInit(): void {
-    const url = localStorage.getItem('theme') || './assets/css/colors/default-dark.css';
-    this.linkTheme?.setAttribute('href', url);
+
   }
 
 }
